Initial Cleanup.
