+++ comp/2026/src/main/java/frc/robot/BuildConstants.java
@@ -7,13 +7,13 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "2026";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 164;
-  public static final String GIT_SHA = "73d0b55bcbe3503f4bc216220fdcb875135e345a";
-  public static final String GIT_DATE = "2026-03-14 19:24:19 EDT";
-  public static final String GIT_BRANCH = "clean-project";
-  public static final String BUILD_DATE = "2026-03-14 19:53:10 EDT";
-  public static final long BUILD_UNIX_TIME = 1773532390178L;
-  public static final int DIRTY = 1;
+  public static final int GIT_REVISION = -1;
+  public static final String GIT_SHA = "UNKNOWN";
+  public static final String GIT_DATE = "UNKNOWN";
+  public static final String GIT_BRANCH = "UNKNOWN";
+  public static final String BUILD_DATE = "2026-03-17 07:56:25 EDT";
+  public static final long BUILD_UNIX_TIME = 1773748585121L;
+  public static final int DIRTY = 69;
 
   private BuildConstants(){}
 }
