+++ comp/2026/src/main/java/frc/robot/RobotContainer.java
@@ -48,7 +48,6 @@ import frc.robot.subsystems.Intake.Intake;
 import frc.robot.subsystems.Intake.IntakeAutoOptions;
 import frc.robot.subsystems.Intake.IntakeIO;
 import frc.robot.subsystems.Intake.IntakeReal;
-import frc.robot.subsystems.Intake.IntakeReal2;
 import frc.robot.subsystems.Shooter.Shooter;
 import frc.robot.subsystems.Shooter.ShooterIO;
 import frc.robot.subsystems.Shooter.ShooterRealQuad;
@@ -120,7 +119,7 @@ public class RobotContainer {
 
                                 m_Hopper = new Hopper(new HopperRealSingle());
                                 m_Hopper.applyState();
-                                intake = new Intake(new IntakeReal2());
+                                intake = new Intake(new IntakeReal());
                                 intake.applyState();
                                 break;
                         case SIM:
@@ -136,7 +135,7 @@ public class RobotContainer {
                                 m_Hopper = new Hopper(new HopperRealSingle());
                                 m_Hopper.applyState();
 
-                                intake = new Intake(new IntakeReal2());
+                                intake = new Intake(new IntakeReal());
                                 intake.applyState();
 
                                 vision = new Vision(drive::addVisionMeasurement,
