+++ comp/2026/src/main/java/frc/robot/subsystems/Intake/IntakeReal.java
@@ -6,6 +6,7 @@ import static edu.wpi.first.units.Units.Rotation;
 import static edu.wpi.first.units.Units.Rotations;
 import static edu.wpi.first.units.Units.RotationsPerSecond;
 import static edu.wpi.first.units.Units.RotationsPerSecondPerSecond;
+import static edu.wpi.first.units.Units.Volt;
 import static edu.wpi.first.units.Units.Volts;
 
 import org.bobcatrobotics.Hardware.Characterization.CharacterizationClosedLoopOutputType;
@@ -34,8 +35,11 @@ public class IntakeReal implements IntakeIO {
   private TalonFX positionMotor;
   public ModuleConfigurator intakePivotConfig;
 
-  private TalonFX velocityMotor;
-  public ModuleConfigurator intakeVelocityConfig;
+  private TalonFX leftVelocityMotor;
+  public ModuleConfigurator leftintakeVelocityConfig;
+
+  private TalonFX rightVelocityMotor;
+  public ModuleConfigurator rightIntakeVelocityConfig;
 
   private TorqueCurrentFOC characterizationRequestTorqueCurrentFOC = new TorqueCurrentFOC(0);
   private VoltageOut characterizationRequestVoltage = new VoltageOut(0);
@@ -49,10 +53,15 @@ public class IntakeReal implements IntakeIO {
   public double intakePivotSetpoint = 0;
   public double intakeVelocitySetpoint = 0;
 
-  private StatusSignal<AngularVelocity> velocityOfIntakeSpeedRPS;
-  private StatusSignal<Current> statorCurrentOfIntakeSpeedAmps;
-  private StatusSignal<Voltage> outputOfIntakeSpeedVolts;
-  private StatusSignal<AngularAcceleration> accelerationOfIntakeSpeed;
+  private StatusSignal<AngularVelocity> leftVelocityOfIntakeSpeedRPS;
+  private StatusSignal<Current> leftStatorCurrentOfIntakeSpeedAmps;
+  private StatusSignal<Voltage> leftOutputOfIntakeSpeedVolts;
+  private StatusSignal<AngularAcceleration> leftAccelerationOfIntakeSpeed;
+
+  private StatusSignal<AngularVelocity> rightVelocityOfIntakeSpeedRPS;
+  private StatusSignal<Current> rightStatorCurrentOfIntakeSpeedAmps;
+  private StatusSignal<Voltage> rightOutputOfIntakeSpeedVolts;
+  private StatusSignal<AngularAcceleration> rightAccelerationOfIntakeSpeed;
 
   private StatusSignal<AngularVelocity> velocityOfIntakePositionRPS;
   private StatusSignal<Current> statorCurrentOfIntakePositionAmps;
@@ -60,8 +69,10 @@ public class IntakeReal implements IntakeIO {
   private StatusSignal<AngularAcceleration> accelerationOfIntakePosition;
 
   // private FindLimit seekLowerRange;
-Gains pivotMotorGains;
-Gains rollerMotorGains;
+  Gains pivotMotorGains;
+  Gains rightRollerMotorGains;
+  Gains leftRollerMotorGains;
+
   public IntakeReal() {
     pivotMotorGains = new Gains.Builder()
         .kP(Constants.IntakeConstants.PivotConstants.kP)
@@ -71,44 +82,78 @@ Gains rollerMotorGains;
         .kV(Constants.IntakeConstants.PivotConstants.kV)
         .kA(Constants.IntakeConstants.PivotConstants.kA)
         .build();
-    rollerMotorGains = new Gains.Builder()
+    rightRollerMotorGains = new Gains.Builder()
         .kP(Constants.IntakeConstants.RightRollerConstants.kP)
         .kI(Constants.IntakeConstants.RightRollerConstants.kI)
         .kD(Constants.IntakeConstants.RightRollerConstants.kD)
         .kS(Constants.IntakeConstants.RightRollerConstants.kS)
         .kV(Constants.IntakeConstants.RightRollerConstants.kV)
         .kA(Constants.IntakeConstants.RightRollerConstants.kA).build();
-    setupRollerMotor(rollerMotorGains);
+
+    leftRollerMotorGains = new Gains.Builder()
+        .kP(Constants.IntakeConstants.LeftRollerConstants.kP)
+        .kI(Constants.IntakeConstants.LeftRollerConstants.kI)
+        .kD(Constants.IntakeConstants.LeftRollerConstants.kD)
+        .kS(Constants.IntakeConstants.LeftRollerConstants.kS)
+        .kV(Constants.IntakeConstants.LeftRollerConstants.kV)
+        .kA(Constants.IntakeConstants.LeftRollerConstants.kA).build();
+    setUpLeftRollerMotor(leftRollerMotorGains);
+    setUpRightRollerMotor(rightRollerMotorGains);
     setupPivotMotor(pivotMotorGains);
 
     // seekLowerRange = new FindLimit(false, positionMotor);
   }
 
-  public void setupRollerMotor(Gains g) {
-    intakeVelocityConfig = new ModuleConfigurator(g.toSlot0Configs(),
+  public void setUpRightRollerMotor(Gains g) {
+    rightIntakeVelocityConfig = new ModuleConfigurator(g.toSlot0Configs(),
         Constants.IntakeConstants.RightRollerConstants.rollerMotorId,
         Constants.IntakeConstants.RightRollerConstants.isInverted,
         Constants.IntakeConstants.RightRollerConstants.isCoast,
         Constants.IntakeConstants.RightRollerConstants.currentLimit,
         Constants.IntakeConstants.RightRollerConstants.peakForwardLimit,
         Constants.IntakeConstants.RightRollerConstants.peakReverseLimit);
-    velocityMotor = new TalonFX(intakeVelocityConfig.getMotorId(), new CANBus("rio"));
-    intakeVelocityConfig.configureMotor(velocityMotor, g);
+    rightVelocityMotor = new TalonFX(rightIntakeVelocityConfig.getMotorId(), new CANBus("rio"));
+    rightIntakeVelocityConfig.configureMotor(rightVelocityMotor, g);
     if (Constants.lowTelemetryMode) {
-      velocityOfIntakeSpeedRPS = velocityMotor.getVelocity();
-      statorCurrentOfIntakeSpeedAmps = velocityMotor.getStatorCurrent();
-      outputOfIntakeSpeedVolts = velocityMotor.getMotorVoltage();
-      intakeVelocityConfig.configureSignals(velocityMotor, 50.0, velocityOfIntakeSpeedRPS,
-          statorCurrentOfIntakeSpeedAmps, outputOfIntakeSpeedVolts);
+      rightVelocityOfIntakeSpeedRPS = rightVelocityMotor.getVelocity();
+      rightStatorCurrentOfIntakeSpeedAmps = rightVelocityMotor.getStatorCurrent();
+      rightOutputOfIntakeSpeedVolts = rightVelocityMotor.getMotorVoltage();
+      rightIntakeVelocityConfig.configureSignals(rightVelocityMotor, 50.0, rightVelocityOfIntakeSpeedRPS,
+          rightStatorCurrentOfIntakeSpeedAmps, rightOutputOfIntakeSpeedVolts);
     } else {
-      velocityOfIntakeSpeedRPS = velocityMotor.getVelocity();
-      statorCurrentOfIntakeSpeedAmps = velocityMotor.getStatorCurrent();
-      outputOfIntakeSpeedVolts = velocityMotor.getMotorVoltage();
-      accelerationOfIntakeSpeed = velocityMotor.getAcceleration();
-      intakeVelocityConfig.configureSignals(velocityMotor, 50.0, velocityOfIntakeSpeedRPS,
-          statorCurrentOfIntakeSpeedAmps, outputOfIntakeSpeedVolts, accelerationOfIntakeSpeed);
+      rightVelocityOfIntakeSpeedRPS = rightVelocityMotor.getVelocity();
+      rightStatorCurrentOfIntakeSpeedAmps = rightVelocityMotor.getStatorCurrent();
+      rightOutputOfIntakeSpeedVolts = rightVelocityMotor.getMotorVoltage();
+      rightAccelerationOfIntakeSpeed = rightVelocityMotor.getAcceleration();
+      rightIntakeVelocityConfig.configureSignals(rightVelocityMotor, 50.0, rightVelocityOfIntakeSpeedRPS,
+          rightStatorCurrentOfIntakeSpeedAmps, rightOutputOfIntakeSpeedVolts, rightAccelerationOfIntakeSpeed);
     }
+  }
 
+  public void setUpLeftRollerMotor(Gains g) {
+    leftintakeVelocityConfig = new ModuleConfigurator(g.toSlot0Configs(),
+        Constants.IntakeConstants.LeftRollerConstants.rollerMotorId,
+        Constants.IntakeConstants.LeftRollerConstants.isInverted,
+        Constants.IntakeConstants.LeftRollerConstants.isCoast,
+        Constants.IntakeConstants.LeftRollerConstants.currentLimit,
+        Constants.IntakeConstants.LeftRollerConstants.peakForwardLimit,
+        Constants.IntakeConstants.LeftRollerConstants.peakReverseLimit);
+    leftVelocityMotor = new TalonFX(leftintakeVelocityConfig.getMotorId(), new CANBus("rio"));
+    leftintakeVelocityConfig.configureMotor(leftVelocityMotor, g);
+    if (Constants.lowTelemetryMode) {
+      leftVelocityOfIntakeSpeedRPS = leftVelocityMotor.getVelocity();
+      leftStatorCurrentOfIntakeSpeedAmps = leftVelocityMotor.getStatorCurrent();
+      leftOutputOfIntakeSpeedVolts = leftVelocityMotor.getMotorVoltage();
+      leftintakeVelocityConfig.configureSignals(leftVelocityMotor, 50.0, leftVelocityOfIntakeSpeedRPS,
+          leftStatorCurrentOfIntakeSpeedAmps, leftOutputOfIntakeSpeedVolts);
+    } else {
+      leftVelocityOfIntakeSpeedRPS = leftVelocityMotor.getVelocity();
+      leftStatorCurrentOfIntakeSpeedAmps = leftVelocityMotor.getStatorCurrent();
+      leftOutputOfIntakeSpeedVolts = leftVelocityMotor.getMotorVoltage();
+      leftAccelerationOfIntakeSpeed = leftVelocityMotor.getAcceleration();
+      leftintakeVelocityConfig.configureSignals(leftVelocityMotor, 50.0, leftVelocityOfIntakeSpeedRPS,
+          leftStatorCurrentOfIntakeSpeedAmps, leftOutputOfIntakeSpeedVolts, leftAccelerationOfIntakeSpeed);
+    }
   }
 
   public void setupPivotMotor(Gains g) {
@@ -148,36 +193,51 @@ Gains rollerMotorGains;
 
   }
 
-  public void highTelemetry(IntakeIOInputs inputs) {
-    BaseStatusSignal.refreshAll(
-        accelerationOfIntakePosition,
-        accelerationOfIntakeSpeed, outputOfIntakePositionVolts, outputOfIntakeSpeedVolts);
-    inputs.accelerationOfIntakePosition = accelerationOfIntakePosition.getValue().in(RotationsPerSecondPerSecond);
-    inputs.rightAccelerationOfIntakeSpeed = accelerationOfIntakeSpeed.getValue().in(RotationsPerSecondPerSecond);
-    inputs.outputOfIntakePositionVolts = outputOfIntakePositionVolts.getValue().in(Volts);
-    inputs.rightOutputOfIntakeSpeedVolts = outputOfIntakeSpeedVolts.getValue()
-        .in(Volts);
-    lowTelemetry(inputs);
-
-  }
-
   public void lowTelemetry(IntakeIOInputs inputs) {
     BaseStatusSignal.refreshAll(
         velocityOfIntakePositionRPS, statorCurrentOfIntakePositionAmps,
-        velocityOfIntakeSpeedRPS, statorCurrentOfIntakeSpeedAmps);
+        rightVelocityOfIntakeSpeedRPS, rightStatorCurrentOfIntakeSpeedAmps,
+        leftVelocityOfIntakeSpeedRPS, leftStatorCurrentOfIntakeSpeedAmps);
+
+    // Position
     inputs.velocityOfIntakePositionRPS = velocityOfIntakePositionRPS.getValue().in(Rotation.per(Minute));
     inputs.statorCurrentOfIntakePositionAmps = statorCurrentOfIntakePositionAmps.getValue().in(Amps);
-    inputs.rightVelocityMotorConnected = velocityMotor.isConnected();
-    inputs.rightVelocityOfIntakeSpeedRPS = velocityOfIntakeSpeedRPS.getValue().in(Rotation.per(Minute));
-    inputs.rightStatorCurrentOfIntakeSpeedAmps = statorCurrentOfIntakeSpeedAmps.getValue().in(Amps);
-    inputs.positionConnected = velocityMotor.isConnected();
+    inputs.positionConnected = positionMotor.isConnected();
     inputs.intakePosition = positionMotor.getPosition().getValueAsDouble();
+
+    // Right Velocity
+    inputs.rightVelocityMotorConnected = rightVelocityMotor.isConnected();
+    inputs.rightVelocityOfIntakeSpeedRPS = rightVelocityOfIntakeSpeedRPS.getValue().in(Rotation.per(Minute));
+    inputs.rightStatorCurrentOfIntakeSpeedAmps = rightStatorCurrentOfIntakeSpeedAmps.getValue().in(Amps);
+
+    // Left Velocity
+    inputs.leftVelocityMotorConnected = leftVelocityMotor.isConnected();
+    inputs.leftVelocityOfIntakeSpeedRPS = leftVelocityOfIntakeSpeedRPS.getValue().in(Rotation.per(Minute));
+    inputs.leftStatorCurrentOfIntakeSpeedAmps = leftStatorCurrentOfIntakeSpeedAmps.getValue().in(Amps);
+
+  }
+
+  public void highTelemetry(IntakeIOInputs inputs) {
+    BaseStatusSignal.refreshAll(
+        accelerationOfIntakePosition, outputOfIntakePositionVolts,
+        leftAccelerationOfIntakeSpeed, rightAccelerationOfIntakeSpeed,
+        leftOutputOfIntakeSpeedVolts, rightOutputOfIntakeSpeedVolts);
+
+    inputs.accelerationOfIntakePosition = accelerationOfIntakePosition.getValue().in(RotationsPerSecondPerSecond);
+    inputs.outputOfIntakePositionVolts = outputOfIntakePositionVolts.getValue().in(Volts);
+    inputs.leftAccelerationOfIntakeSpeed = leftAccelerationOfIntakeSpeed.getValue().in(RotationsPerSecondPerSecond);
+    inputs.rightAccelerationOfIntakeSpeed = rightAccelerationOfIntakeSpeed.getValue().in(RotationsPerSecondPerSecond);
+    inputs.leftOutputOfIntakeSpeedVolts = leftOutputOfIntakeSpeedVolts.getValue().in(Volts);
+    inputs.rightOutputOfIntakeSpeedVolts = rightOutputOfIntakeSpeedVolts.getValue().in(Volts);
+
+    lowTelemetry(inputs);
   }
 
   public void setVelocity(double velocity) {
     intakeVelocitySetpoint = velocity;
     // velocityMotor.setControl(requestVelocity.withVelocity(velocity).withFeedForward(0.6));
-    velocityMotor.set(1);
+    rightVelocityMotor.set(1);
+    leftVelocityMotor.set(1);
   }
 
   public void setVelocity(IntakeState desiredState) {
@@ -190,7 +250,6 @@ Gains rollerMotorGains;
 
   public void setPosition(double pos) {
     intakePivotSetpoint = pos;
-
     positionMotor.setControl(requestPositionVoltage.withPosition(pos).withFeedForward(.8));
 
   }
@@ -200,8 +259,12 @@ Gains rollerMotorGains;
     positionMotor.setControl(requestPositionVoltage.withPosition(0).withFeedForward(-.7));
   }
 
-  public double getVelocity() {
-    return velocityMotor.getVelocity().getValueAsDouble();
+  public double getRightVelocity() {
+    return rightVelocityMotor.getVelocity().getValueAsDouble();
+  }
+
+  public double getLeftVelocity() {
+    return leftVelocityMotor.getVelocity().getValueAsDouble();
   }
 
   public void resetEncoder() {
@@ -215,23 +278,31 @@ Gains rollerMotorGains;
   }
 
   public void stopRollerWheel() {
-    stopRightRollerWheel();
     stopLeftRollerWheel();
+    stopRightRollerWheel();
   }
-    public void stopLeftRollerWheel() {
-  }
-    public void stopRightRollerWheel() {
+
+  public void stopLeftRollerWheel() {
     intakeVelocitySetpoint = 0.0;
-    velocityMotor.stopMotor();
+    leftVelocityMotor.stopMotor();
+    rightVelocityMotor.stopMotor();
+  }
+
+  public void stopRightRollerWheel() {
+    intakeVelocitySetpoint = 0.0;
+    leftVelocityMotor.stopMotor();
+    rightVelocityMotor.stopMotor();
   }
 
   public void stopPivotMotor() {
     intakePivotSetpoint = 0.0;
-    velocityMotor.stopMotor();
+    positionMotor.stopMotor();
   }
 
   public void stopBottom() {
   }
+
+  
 
   @Override
   public void periodic() {
@@ -243,7 +314,11 @@ Gains rollerMotorGains;
 
   /* Characterization */
   public void runCharacterization_IntakeVelocity(double output) {
-    velocityMotor.setControl(switch (CharacterizationClosedLoopOutputType.Voltage) {
+    rightVelocityMotor.setControl(switch (CharacterizationClosedLoopOutputType.Voltage) {
+      case Voltage -> characterizationRequestVoltage.withOutput(output);
+      case TorqueCurrentFOC -> characterizationRequestTorqueCurrentFOC.withOutput(output);
+    });
+    leftVelocityMotor.setControl(switch (CharacterizationClosedLoopOutputType.Voltage) {
       case Voltage -> characterizationRequestVoltage.withOutput(output);
       case TorqueCurrentFOC -> characterizationRequestTorqueCurrentFOC.withOutput(output);
     });
@@ -251,7 +326,7 @@ Gains rollerMotorGains;
 
   /* Characterization */
   public void runCharacterization_IntakePosition(double output) {
-    velocityMotor.setControl(switch (CharacterizationClosedLoopOutputType.Voltage) {
+    positionMotor.setControl(switch (CharacterizationClosedLoopOutputType.Voltage) {
       case Voltage -> characterizationRequestVoltage.withOutput(output);
       case TorqueCurrentFOC -> characterizationRequestTorqueCurrentFOC.withOutput(output);
     });
@@ -259,7 +334,8 @@ Gains rollerMotorGains;
 
   /** Returns the module velocity in rotations/sec (Phoenix native units). */
   public double getFFCharacterizationVelocity_Intake() {
-    double avg = (velocityMotor.getVelocity().getValue().in(RotationsPerSecond)) / 1;
+    double avg = (leftVelocityMotor.getVelocity().getValue().in(RotationsPerSecond)
+        + rightVelocityMotor.getVelocity().getValue().in(RotationsPerSecond)) / 2;
     return avg;
   }
 
@@ -279,7 +355,8 @@ Gains rollerMotorGains;
         Constants.IntakeConstants.PivotConstants.peakReverseLimit);
     intakePivotConfig.configureMotor(positionMotor, pivotMotorGains);
   }
-    public void setNeturalBrake() {
+
+  public void setNeturalBrake() {
     intakePivotConfig = new ModuleConfigurator(pivotMotorGains.toSlot0Configs(),
         Constants.IntakeConstants.PivotConstants.pivotMotorId,
         Constants.IntakeConstants.PivotConstants.isInverted,
